refactor: stabilize API foundation with Scalar, extensions, and namespace fixes
- Add Scalar.AspNetCore for OpenAPI UI and extract presentation setup into Extensions
- Remove boilerplate WeatherForecast file
- Fix GetHomeById namespace consistency across query and handler
