+++ src/StuffTracker.API/Extensions/WebApplicationBuilderExtensions.cs
@@ -1,0 +1,20 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+using Microsoft.AspNetCore.OpenApi;
+
+namespace StuffTracker.API.Extensions;
+
+public static class WebApplicationBuilderExtensions
+{
+    public static void AddPresentation(this WebApplicationBuilder builder)
+    {
+        builder.Services.AddOpenApi();
+
+    builder.Services.AddControllers();
+    }
+}
+
+
