+++ src/StuffTracker.Application/Locations/Queries/GetHomeById/GetHomeByIdQuery.cs
@@ -2,6 +2,6 @@
 
 using StuffTracker.Application.Locations.Dtos;
 
-namespace StuffTracker.Application.Locations.Queries.GetHome;
+namespace StuffTracker.Application.Locations.Queries.GetHomeById;
 
 public record class GetHomeByIdQuery(Guid Id) : IRequest<HomeDto>;
